Factored available_options into a separate class for each command.
