+++ lib/branch_io_cli/configuration.rb
@@ -2,6 +2,9 @@ require "branch_io_cli/configuration/configuration"
 require "branch_io_cli/configuration/option"
 require "branch_io_cli/configuration/option_wrapper"
 require "branch_io_cli/configuration/report_configuration"
+require "branch_io_cli/configuration/report_options"
 require "branch_io_cli/configuration/setup_configuration"
+require "branch_io_cli/configuration/setup_options"
 require "branch_io_cli/configuration/validate_configuration"
+require "branch_io_cli/configuration/validate_options"
 require "branch_io_cli/configuration/xcode_settings"
